requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.1
streamlit==1.34.0
pandas==2.2.2
plotly==5.22.0
apscheduler==3.10.4
